organized Modal. Moved to /common.
