+++ web/src/components/DisplayComponent.js
@@ -1,8 +1,7 @@
 import React, { Component, cloneElement, Children, useContext } from 'react';
 import { withRouter } from 'react-router-dom';
 
-import { Input, Button, Table, TableRowWithButtons, TableRowWithCheckbox } from './common';
-import Modal from './Modal';
+import { Input, Button, Table, TableRowWithButtons, TableRowWithCheckbox, Modal } from './common';
 import DialogModal from './DialogModal';
 import { editPathWithId, itemPathWithId } from './frontendBaseRoutes';
 import Pager from './Pager';
